Add allow list for tool categories (#100)
Right now it is only possible to disable categories of tools which is
tedious for users who want to only use one or two categories, and means
that new tools added in the future will automatically show up. Add the
ability to also set an allow-list of tools that can then be overridden
using the --disable-<category> flags.