Support X-Access-Token for Grafana Cloud (#107)
* Support X-Access-Token for Grafana Cloud

In Grafana Cloud it is possible to have a plugin submit requests on
behalf of the calling user by setting the X-Access-Token header instead
of the standard Authorization header. This PR will preferentially look
for the presence of an access token and use that before falling back to
an API key like normal.

* Add user token for on behalf of auth